Rename decoration -> decorator
