+++ src/flask_request_check/__init__.py
@@ -1,1 +1,1 @@
-from .decoration import validate
+from .decorator import validate
